SS-12: One client per one ticket rule

+++ src/main/kotlin/com/kotlin/skiservice/dto/rental/RentalOrderRequest.kt
@@ -1,7 +1,7 @@
 package com.kotlin.skiservice.dto.rental
 
 data class RentalOrderRequest(
-    val ticketNumber: Long,
+    val ticketNumber: Int,
     val documentType: String,
     val documentNumber: String,
     val cellNumber: String
